complete restructure, moved to the squirtgun platform model
