+++ pronouns-bukkit/build.gradle.kts
@@ -1,19 +1,25 @@
 import com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar
+import org.apache.tools.ant.filters.ReplaceTokens
 
 plugins {
     id("com.github.johnrengelman.shadow") version "7.0.0"
 }
 
+repositories {
+    maven("https://hub.spigotmc.org/nexus/content/repositories/snapshots/")
+}
+
 dependencies {
     implementation(project(":pronouns-api"))
-    implementation("me.lucyy:squirtgun-api:2.0.0-pre3")
-    implementation("me.lucyy:squirtgun-bukkit:2.0.0-pre3")
-    implementation("me.lucyy:squirtgun-commands:2.0.0-pre3")
+    implementation(project(":pronouns-squirtgun"))
+    implementation("me.lucyy:squirtgun-api:2.0.0-pre4-SNAPSHOT")
+    implementation("me.lucyy:squirtgun-bukkit:2.0.0-pre4-SNAPSHOT")
+    implementation("me.lucyy:squirtgun-commands:2.0.0-pre4-SNAPSHOT")
     implementation("com.zaxxer:HikariCP:4.0.2")
     implementation("org.bstats:bstats-bukkit:2.2.1")
-    compileOnly("com.destroystokyo.paper:paper-api:1.16.5-R0.1-SNAPSHOT")
-    compileOnly("net.kyori:adventure-api:4.7.0")
-    compileOnly("net.kyori:examination-api:1.1.0")
+
+    compileOnly("org.spigotmc:spigot-api:1.17-R0.1-SNAPSHOT")
+    compileOnly("net.kyori:adventure-api:4.8.1")
     compileOnly("com.google.guava:guava:30.1.1-jre")
     compileOnly("org.jetbrains:annotations:13.0")
 }
@@ -31,15 +37,24 @@ tasks {
             exclude(project(":pronouns-api"))
         }
 
+        dependencies {
+            exclude(dependency("org.checkerframework:.*:.*"))
+            exclude(dependency("com.google..*:.*:.*"))
+        }
         // slf4j is transitive from hikari
         relocate("org.slf4j", "me.lucyy.pronouns.deps.slf4j")
         relocate("me.lucyy.squirtgun", "me.lucyy.pronouns.deps.squirtgun")
+        relocate("net.kyori", "me.lucyy.pronouns.deps.kyori")
         relocate("org.bstats", "me.lucyy.pronouns.deps.bstats")
         relocate("com.zaxxer.hikari", "me.lucyy.pronouns.deps.hikari")
     }
 
     named("build") {
         dependsOn(shadowTask)
+    }
+
+    withType<ProcessResources> {
+        filter<ReplaceTokens>("tokens" to mapOf("version" to project.version.toString()))
     }
 }
