+++ pronouns-bukkit/src/main/resources/plugin.yml
@@ -1,6 +1,6 @@
 name: ProNouns
-version: ${project.version}
-main: me.lucyy.pronouns.ProNouns
+version: @version@
+main: me.lucyy.pronouns.bukkit.ProNounsBukkit
 api-version: 1.16
 authors: [ __lucyy ]
 description: Putting the "pro" in "pronouns".
