+++ pronouns-squirtgun/src/main/java/me/lucyy/pronouns/command/ShowPronounsNode.java
@@ -26,9 +26,9 @@ import me.lucyy.squirtgun.command.argument.OnlinePlayerArgument;
 import me.lucyy.squirtgun.command.context.CommandContext;
 import me.lucyy.squirtgun.command.node.CommandNode;
 import me.lucyy.squirtgun.format.FormatProvider;
-import me.lucyy.squirtgun.platform.PermissionHolder;
 import me.lucyy.squirtgun.platform.Platform;
-import me.lucyy.squirtgun.platform.SquirtgunPlayer;
+import me.lucyy.squirtgun.platform.audience.PermissionHolder;
+import me.lucyy.squirtgun.platform.audience.SquirtgunPlayer;
 import net.kyori.adventure.text.Component;
 import org.jetbrains.annotations.NotNull;
