+++ pronouns-squirtgun/src/main/java/me/lucyy/pronouns/config/SqlInfoContainer.java
