+++ pronouns-squirtgun/src/main/java/me/lucyy/pronouns/storage/MysqlConnectionException.java
