+++ settings.gradle.kts
@@ -4,5 +4,6 @@
 
 rootProject.name = "pronouns"
 include(":pronouns-api")
+include(":pronouns-squirtgun")
 include(":pronouns-papi")
 include(":pronouns-bukkit")
